move dsp code, do some ui fixes

+++ src/ui/CR42YWFView.h
@@ -51,7 +51,7 @@ public:
 
 	void setShowAll(bool showAll);
 
-	void setSelectedWaveform(int selectedWaveform);
+	void setSelectedWaveform(float selectedWaveform);
 
 	void setData(WavetableEditData* data);
